- aint no sunshine when she's gone - no ya a 7abibi there is sun w shine w kol 7aga

+++ lib/c_controllers/i_flyer_controllers/header_controller.dart
@@ -159,7 +159,7 @@ void _triggerHeaderPageOpacity({
 }) {
 
   final ActiveFlyerProvider _activeFlyerProvider = Provider.of<ActiveFlyerProvider>(context, listen: false);
-  final int _headerPageOpacity = _activeFlyerProvider.headerPageOpacity;
+  final double _headerPageOpacity = _activeFlyerProvider.headerPageOpacity;
 
   blog('_headerPageOpacity = $_headerPageOpacity');
